Some aarch64 fixes

+++ src/gauche/priv/arith_aarch64.h
@@ -40,12 +40,12 @@
 /* cmp %2, #1 sets C = (c >= 1); since c in {0,1}, this makes C = c.
    adcs computes x + y + C_in and updates NZCV.
    adc harvests the resulting carry into the new c. */
-#define UADD(r, c, x, y)                        \
-    asm("cmp %2, #1;"                           \
-        "adcs %0, %3, %4;"                      \
-        "adc %1, xzr, xzr;"                     \
-        : "=&r" (r), "=&r" (c)                  \
-        : "1" (c), "r" (x), "r" (y)             \
+#define UADD(r, c, x, y)                                \
+    asm("cmp %2, #1;"                                   \
+        "adcs %0, %3, %4;"                              \
+        "adc %1, xzr, xzr;"                             \
+        : "=&r" (r), "=&r" (c)                          \
+        : "1" (c), "r" ((u_long)x), "r" ((u_long)y)     \
         : "cc")
 
 #endif //SCM_ENABLE_ALL_ARITH_ASMS
@@ -63,7 +63,7 @@
     asm("adds %0, %2, %3;"                      \
         "cset %1, cs;"                          \
         : "=&r" (r), "=&r" (v)                  \
-        : "r" (x), "r" (y)                      \
+        : "r" ((u_long)x), "r" ((u_long)y)      \
         : "cc")
 
 #endif //SCM_ENABLE_ALL_ARITH_ASMS
@@ -85,7 +85,7 @@
         "cset %1, vs;"                          \
         "cneg %1, %1, pl;"                      \
         : "=&r" (r), "=&r" (v)                  \
-        : "r" (x), "r" (y)                      \
+        : "r" ((long)x), "r" ((long)y)          \
         : "cc")
 
 /*-----------------------------------------------------------------
@@ -99,12 +99,12 @@
    'cmp xzr, %2' computes 0 - c: C = 1 iff 0 >= c (unsigned), i.e., c=0.
    sbcs then does x - y - (1 - C) = x - y - c.
    cset %1, cc yields 1 iff borrow (C_out clear). */
-#define USUB(r, c, x, y)                        \
-    asm("cmp xzr, %2;"                          \
-        "sbcs %0, %3, %4;"                      \
-        "cset %1, cc;"                          \
-        : "=&r" (r), "=&r" (c)                  \
-        : "1" (c), "r" (x), "r" (y)             \
+#define USUB(r, c, x, y)                                \
+    asm("cmp xzr, %2;"                                  \
+        "sbcs %0, %3, %4;"                              \
+        "cset %1, cc;"                                  \
+        : "=&r" (r), "=&r" (c)                          \
+        : "1" (c), "r" ((u_long)x), "r" ((u_long)y)     \
         : "cc")
 
 /*-----------------------------------------------------------------
@@ -120,7 +120,7 @@
     asm("subs %0, %2, %3;"                      \
         "cset %1, cc;"                          \
         : "=&r" (r), "=&r" (v)                  \
-        : "r" (x), "r" (y)                      \
+        : "r" ((u_long)x), "r" ((u_long)y)      \
         : "cc")
 
 #endif // SCM_ENABLE_ALL_ARITH_ASMS
@@ -163,8 +163,7 @@
 
 /*-----------------------------------------------------------------
  * UMULOV(r, v, x, y)      unsigned word multiply with overflow check
- *  u_long : r, x, y
- *  int : v
+ *  u_long : r, x, y, v
  *  if x * y overflows, v = 1
  *  else r <- x * y, v = 0
  */
@@ -176,13 +175,12 @@
         "cmp %1, #0;"                           \
         "cset %1, ne;"                          \
         : "=&r" (r), "=&r" (v)                  \
-        : "r" (x), "r" (y)                      \
+        : "r" ((u_long)x), "r" ((u_long)y)      \
         : "cc")
 
 /*-----------------------------------------------------------------
  * SMULOV(r, v, x, y)      signed word multiply with overflow check
- *  long : r, x, y
- *  int : v
+ *  long : r, x, y, v
  *  if x * y overflows, v = 1 or -1 depending on the sign of the result
  *  else r <- x * y, v = 0
  */
@@ -205,7 +203,7 @@
         "0: mov %1, #0;"                        \
         "2:"                                    \
         : "=&r" (r), "=&r" (v)                  \
-        : "r" (x), "r" (y)                      \
+        : "r" ((u_long)x), "r" ((u_long)y)      \
         : "cc")
 
 #endif //SCM_ENABLE_ALL_ARITH_ASMS
